Fixed loss of precision (FloatingPointException) when calculating duration

+++ independent_method.h
@@ -19,8 +19,8 @@ public:
     using AbstractMethod::AbstractMethod;
 
     void work(int thread_index, const vector<tuple<uint64_t, uint64_t>>& data, int start_index, int bucket_size);
-    int thread_work(const vector<tuple<uint64_t, uint64_t>>& data);
-    int thread_work_affinity(const vector<tuple<uint64_t, uint64_t>>& data);
+    double thread_work(const vector<tuple<uint64_t, uint64_t>>& data);
+    double thread_work_affinity(const vector<tuple<uint64_t, uint64_t>>& data);
     void print_buffers_everything();
     void print_buffers_partition_entries();
     void print_buffers_partition_statistics();
